updated the code

+++ src/main/java/com/target_india/digitize_time_table/repository/CourseDao.java
@@ -2,6 +2,7 @@ package com.target_india.digitize_time_table.repository;
 
 import com.target_india.digitize_time_table.config.DbConnection;
 import com.target_india.digitize_time_table.exceptions.ResourceAlreadyExistException;
+import com.target_india.digitize_time_table.exceptions.ResourceNotFoundException;
 import com.target_india.digitize_time_table.model.Course;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -28,11 +29,11 @@ public class CourseDao {
 
     public ResultSet findCourseByClassId(int classId){
         ResultSet resultSet=null;
+        String query = "select course_table.course_id,course_table.course_name,instructor_table.instructor_name " +
+                "from course_table " +
+                "inner join instructor_table on course_table.instructor_id = instructor_table.instructor_id " +
+                "where course_table.class_id = ?";
         try {
-            String query = "select course_table.course_id,course_table.course_name,instructor_table.instructor_name " +
-                    "from course_table " +
-                    "inner join instructor_table on course_table.instructor_id = instructor_table.instructor_id " +
-                    "where course_table.class_id = ?";
             PreparedStatement preparedStatement = connection.prepareStatement(query);
             preparedStatement.setInt(1, classId);
             resultSet = preparedStatement.executeQuery();
@@ -45,8 +46,8 @@ public class CourseDao {
 
     public ResultSet findCourseById(int id) {
         ResultSet resultSet = null;
+        String query = "select * from course_table where course_id = ?";
         try {
-            String query = "select * from course_table where course_id = ?";
             PreparedStatement preparedStatement = connection.prepareStatement(query);
             preparedStatement.setInt(1, id);
             resultSet = preparedStatement.executeQuery();
@@ -58,9 +59,8 @@ public class CourseDao {
     }
 
     public void addCourse(Course course) {
+        String query = "insert into course_table(course_id,course_name,class_id,instructor_id) values(?,?,?,?)";
         try {
-            String query = "insert into course_table(course_id,course_name,class_id,instructor_id) values(?,?,?,?)";
-
             PreparedStatement preparedStatement = connection.prepareStatement(query);
             preparedStatement.setInt(1,course.getCourseId());
             preparedStatement.setString(2,course.getCourseName());
@@ -75,7 +75,8 @@ public class CourseDao {
 
     public String updateCourse(int id, String name, int classId, int instructorId) {
         Statement statement;
-        try(ResultSet resultSet = findCourseById(id)){
+        ResultSet resultSet = findCourseById(id);
+        try{
             if(resultSet.next()){
                 String query="update course_table " +
                         "set course_name='"+name+"', class_id='"+classId+"', instructor_id="+instructorId+" " +
@@ -85,14 +86,12 @@ public class CourseDao {
                 return "Successfully updated course data whose id = "+ id;
             }
             else{
-                return null;
+                throw new ResourceNotFoundException("No course found with ID"+id+"to update");
             }
 
         }
         catch(Exception exception) {
-            logger.error(String.valueOf(exception));
-
-            throw new ResourceAlreadyExistException(exception.getMessage());
+            throw new ResourceNotFoundException(String.valueOf(exception));
         }
     }
 
@@ -112,26 +111,28 @@ public class CourseDao {
                 return "Successfully updated student data whose id = "+ course.getCourseId();
             }
             else{
-                return null;
+                throw new ResourceNotFoundException("No course found with ID"+course.getCourseId()+"to update");
             }
 
         }
-        catch(Exception exception){
-            logger.error(String.valueOf(exception));
-            throw new ResourceAlreadyExistException(exception.getMessage());
+        catch(SQLException exception){
+            throw new ResourceNotFoundException(String.valueOf(exception));
         }
     }
 
-    public int deleteCourseById(int id) {
+    public String deleteCourseById(int id) {
+        ResultSet resultset = findCourseById(id);
         try {
-            String query = "Delete from course_table where course_id = " + id;
-
-            Statement statement = connection.createStatement();
-            statement.executeUpdate(query);
-            return 1;
+            if (resultset.next()) {
+                String query = "Delete from course_table where course_id = " + id;
+                Statement statement = connection.createStatement();
+                statement.executeUpdate(query);
+                return "Successfully deleted course data whose id = " + id;
+            }
+            throw new ResourceNotFoundException("No course found with ID" + id + "to delete");
         }
         catch (Exception exception){
-            return 0;
+            throw new ResourceNotFoundException(String.valueOf(exception));
         }
     }
 }
