+++ src/main/java/com/target_india/digitize_time_table/repository/StudentDao.java
@@ -2,6 +2,8 @@ package com.target_india.digitize_time_table.repository;
 
 import com.target_india.digitize_time_table.config.DbConnection;
 
+import com.target_india.digitize_time_table.exceptions.ResourceAlreadyExistException;
+import com.target_india.digitize_time_table.exceptions.ResourceNotFoundException;
 import com.target_india.digitize_time_table.model.Student;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -28,8 +30,8 @@ public class StudentDao {
 
     public ResultSet findStudentById(int id) {
         ResultSet resultSet = null;
+        String query = "select * from student_table where student_id = ?";
         try {
-            String query = "select * from student_table where student_id = ?";
             PreparedStatement preparedStatement = connection.prepareStatement(query);
             preparedStatement.setInt(1, id);
             resultSet = preparedStatement.executeQuery();
@@ -55,14 +57,13 @@ public class StudentDao {
 
 
     public void addStudent(Student student) {
+        String query = "insert into student_table(student_name,class_id,student_contact) values(?,?,?)";
         try {
-            String query = "insert into student_table(student_id,student_name,class_id,student_contact) values(?,?,?,?)";
 
             PreparedStatement preparedStatement = connection.prepareStatement(query);
-            preparedStatement.setInt(1,student.getStudentId());
-            preparedStatement.setString(2,student.getStudentName());
-            preparedStatement.setInt(3,student.getClassId());
-            preparedStatement.setString(4,student.getStudentContact());
+            preparedStatement.setString(1,student.getStudentName());
+            preparedStatement.setInt(2,student.getClassId());
+            preparedStatement.setString(3,student.getStudentContact());
             preparedStatement.executeUpdate();
         }
         catch(SQLException exception){
@@ -71,8 +72,8 @@ public class StudentDao {
     }
     public String updateStudent(Student student) {
         PreparedStatement preparedStatement;
+        ResultSet resultset = findStudentById(student.getStudentId());
         try{
-            ResultSet resultset = findStudentById(student.getStudentId());
             if(resultset.next()){
                 String query="update student_table " +
                         "set student_name=?, class_id = ?, student_contact=? " +
@@ -86,7 +87,7 @@ public class StudentDao {
                 return "Successfully updated student data whose id = "+ student.getStudentId();
             }
             else{
-                return null;
+                throw new ResourceNotFoundException("No student found with ID"+student.getStudentId()+"to update");
             }
 
         }
@@ -99,9 +100,12 @@ public class StudentDao {
 
     public String updateStudent(int id, String name, int classId, String contact) {
         Statement statement;
+        ResultSet resultSet = findStudentById(id);
         try{
-            ResultSet resultSet = findStudentById(id);
             if(resultSet.next()){
+                if(resultSet.getString("student_contact").equals(contact)){
+                    throw new ResourceAlreadyExistException("This contact already exists!");
+                }
                 String query="update student_table " +
                         "set student_name='"+name+"', class_id='"+classId+"', student_contact="+contact+" " +
                         "where student_id="+id+" ";
@@ -109,28 +113,27 @@ public class StudentDao {
                 statement.executeUpdate(query);
                 return "Successfully updated student data whose id = "+ id;
             }
-            else{
-                return null;
-            }
+            throw new ResourceNotFoundException("No student found with ID"+id+"to update");
 
         }
-        catch(Exception exception){
-            logger.error(String.valueOf(exception));
+        catch(SQLException exception){
+            throw new ResourceNotFoundException(String.valueOf(exception));
         }
-        return null;
     }
 
-    public int deleteStudentById(int id) {
+    public String deleteStudentById(int id) {
+        ResultSet resultset = findStudentById(id);
         try {
-            String query = "Delete from student_table where student_id = " + id;
-
-            Statement statement = connection.createStatement();
-            statement.executeUpdate(query);
-            return 1;
+            if (resultset.next()) {
+                String query = "Delete from student_table where student_id = " + id;
+                Statement statement = connection.createStatement();
+                statement.executeUpdate(query);
+                return "Successfully deleted student data whose id = " + id;
+            }
+            throw new ResourceNotFoundException("No student found with ID"+id+"to delete");
         }
-        catch (Exception exception){
-            return 0;
+        catch (SQLException exception){
+            throw new ResourceNotFoundException(String.valueOf(exception));
         }
     }
-
 }
